move model of product list item context and improve unit tests
