+++ projects/storefrontlib/src/cms-components/product/index.ts
@@ -6,7 +6,7 @@ export * from './product-images/product-images.module';
 export * from './product-intro/product-intro.component';
 export * from './product-intro/product-intro.module';
 export * from './product-list/index';
-export * from './product-list/model/product-list-item.context';
+export * from './product-list/model/product-list-item-context.model';
 export * from './product-outlets.model';
 export * from './product-summary/product-summary.component';
 export * from './product-summary/product-summary.module';
