+++ projects/storefrontlib/src/cms-components/product/product-list/index.ts
@@ -1,4 +1,5 @@
 export * from './container/index';
+export * from './model/index';
 export * from './product-facet-navigation/index';
 export * from './product-grid-item/product-grid-item.component';
 export * from './product-list-item/product-list-item.component';
