+++ projects/storefrontlib/src/cms-components/product/product-list/model/index.ts
@@ -1,0 +1,2 @@
+export * from './product-list-item-context-source.model';
+export * from './product-list-item-context.model';
