+++ projects/storefrontlib/src/cms-components/product/product-list/model/product-list-item-context-source.model.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { Product } from '@spartacus/core';
-import { ProductListItemContextSource } from './product-list-item.context';
+import { ProductListItemContextSource } from './product-list-item-context-source.model';
 
 describe('ProductListItemContextSource', () => {
   let contextSource: ProductListItemContextSource;
