+++ projects/storefrontlib/src/cms-components/product/product-list/model/product-list-item-context.model.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@angular/core';
+import { Product } from '@spartacus/core';
+import { Observable } from 'rxjs';
+
+/**
+ * Context for `ProductListItemComponent`.
+ */
+@Injectable()
+export abstract class ProductListItemContext {
+  readonly product$: Observable<Product>;
+}
